requirements.txt
pandas
yfinance
openpyxl
python-dotenv